project tree simplification and qol improvements
