+++ deepecohab/__init__.py
@@ -1,16 +1,13 @@
-from deepecohab.antenna_analysis.activity import (
+from deepecohab.analysis.antenna_analysis import (
 	calculate_activity as calculate_activity,
 	calculate_cage_occupancy as calculate_cage_occupancy,
-)
-from deepecohab.antenna_analysis.chasings import (
 	calculate_chasings as calculate_chasings,
 	calculate_ranking as calculate_ranking,
-)
-from deepecohab.antenna_analysis.incohort_sociability import (
 	calculate_incohort_sociability as calculate_incohort_sociability,
 	calculate_pairwise_meetings as calculate_pairwise_meetings,
 	calculate_time_alone as calculate_time_alone,
 )
+
 from deepecohab.core.create_data_structure import (
 	get_ecohab_data_structure as get_ecohab_data_structure,
 )
