+++ deepecohab/analysis/__init__.py
@@ -1,0 +1,9 @@
+from deepecohab.analysis.antenna_analysis import (
+	calculate_activity,
+	calculate_cage_occupancy,
+	calculate_chasings,
+	calculate_ranking,
+	calculate_time_alone,
+	calculate_pairwise_meetings,
+	calculate_incohort_sociability,
+)
