--- deepecohab/antenna_analysis/__init__.py
@@ -1,6 +1,0 @@
-from deepecohab.antenna_analysis.activity import calculate_activity, calculate_cage_occupancy
-from deepecohab.antenna_analysis.chasings import calculate_chasings, calculate_ranking
-from deepecohab.antenna_analysis.incohort_sociability import calculate_time_alone, calculate_pairwise_meetings, calculate_incohort_sociability
-
-
-# NOTE: Can be dropped after reguistry is implemented in the analysis page
