+++ deepecohab/utils/auxfun_dashboard.py
@@ -391,7 +391,7 @@ def generate_download_block() -> dbc.Modal:
 
 	return modal
 
-def generate_sidebar(icon_map, page_registry):
+def generate_sidebar(icon_map: dict[str, str], page_registry: dict[str, str]):
 	return html.Div(
 		[
 			html.Div("MENU", className="sidebar-label"),
